rename pipeline to pipe
