+++ src/__spec__/pipe.spec.ts
@@ -1,9 +1,9 @@
 import assert from 'static-type-assert'
-import { pipeline, range, map, filter } from '../index'
+import { pipe, range, map, filter } from '../index'
 
 assert<
   IterableIterator<number>
->(pipeline(
+>(pipe(
   range(),
   filter(x => {
     assert<number>(x)
