+++ src/index.mjs
@@ -28,7 +28,7 @@ export { default as merge } from './merge'
 export { mergeByComparison, mergeByChance, mergeByPosition } from './merge'
 export { default as partition } from './partition'
 export { default as permutations } from './permutations'
-export { default as pipeline } from './pipeline'
+export { default as pipe } from './pipe'
 export { default as product } from './product'
 export { default as range } from './range'
 export { default as reduce } from './reduce'
