+++ src/pipe.d.ts
@@ -1,0 +1,2 @@
+import { pipe } from './index'
+export default pipe
