+++ src/pipe.mjs
@@ -1,0 +1,3 @@
+export default function pipe (value, ...fns) {
+  return fns.reduce((value, fn) => fn(value), value)
+}
