--- src/pipeline.d.ts
@@ -1,2 +1,0 @@
-import { pipeline } from './index'
-export default pipeline
