--- src/pipeline.mjs
@@ -1,3 +1,0 @@
-export default function pipeline (iterable, ...fns) {
-  return fns.reduce((iterable, fn) => fn(iterable), iterable)
-}
